requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
pandas==2.1.4
torch==2.1.1
torchvision==0.16.1
transformers==4.36.0
sentencepiece==0.1.99
protobuf==4.25.1
textblob==0.17.1
